requirements.txt
beautifulsoup4==4.8.1
certifi==2019.9.11
chardet==3.0.4
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==5.2.1
future==0.18.2
idna==2.8
jieba3k==0.35.1
joblib==0.14.0
lxml==4.4.1
newspaper3k==0.2.8
nltk==3.4.5
numpy==1.17.3
oauthlib==3.1.0
Pillow==6.2.1
psycopg2-binary==2.8.4
python-dateutil==2.8.1
python-twitter==3.5
PyYAML==5.1.2
requests==2.22.0
requests-file==1.4.3
requests-oauthlib==1.3.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.13.0
soupsieve==1.9.5
tinysegmenter==0.3
tldextract==2.2.2
urllib3==1.25.6
